Move FiLM config to v3 with shared _pretrain_base and proper init docs
- Create v3/vit5_small_pretrain_hyena_cls_row_gated_film_ema.py using
  get_base_config() and make_block_cfg() from _pretrain_base
- Replace small_init/wang_init with trunc_normal_init_factory(std=0.02)
- Add **kwargs to make_block_cfg for FiLM-specific block params
  (register_pooling_cfg, num_registers)
- Remove old v2/film/ config
- Document identity-modulation init pattern in KernelFiLMGenerator

Made-with: Cursor
